feat(laravel-insights): Redirect feedback
