+++ static/app/views/insights/pages/domainViewHeader.tsx
@@ -19,6 +19,7 @@ import {
   type RoutableModuleNames,
   useModuleURLBuilder,
 } from 'sentry/views/insights/common/utils/useModuleURL';
+import {useIsLaravelPageActive} from 'sentry/views/insights/pages/backend/laravel/utils';
 import {OVERVIEW_PAGE_TITLE} from 'sentry/views/insights/pages/settings';
 import {
   isModuleConsideredNew,
@@ -56,6 +57,7 @@ export function DomainViewHeader({
   const organization = useOrganization();
   const location = useLocation();
   const moduleURLBuilder = useModuleURLBuilder();
+  const isLaravelPageActive = useIsLaravelPageActive();
 
   const crumbs: Crumb[] = [
     {
@@ -103,7 +105,18 @@ export function DomainViewHeader({
         </Layout.HeaderContent>
         <Layout.HeaderActions>
           <ButtonBar gap={1}>
-            <FeedbackWidgetButton />
+            <FeedbackWidgetButton
+              optionOverrides={
+                isLaravelPageActive
+                  ? {
+                      tags: {
+                        ['feedback.source']: 'laravel-insights',
+                        ['feedback.owner']: 'telemetry-experience',
+                      },
+                    }
+                  : undefined
+              }
+            />
             {additonalHeaderActions}
           </ButtonBar>
         </Layout.HeaderActions>
